LayoutTests/imported/w3c/web-platform-tests/css/css-cascade/scope-import-multiple.tentative-expected.txt

Inside
Inside
Outside

FAIL A stylesheet may be imported multiple times, and scoped differently assert_equals: expected "1" but got ""
PASS Scope-imported rule does not apply outside of scope
